// packages/wallet-sdk/src/generated/types/delegateCreateOrMutateArgs.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getDiscriminatedUnionDecoder,
  getDiscriminatedUnionEncoder,
  getStructDecoder,
  getStructEncoder,
  getTupleDecoder,
  getTupleEncoder,
  type Codec,
  type Decoder,
  type Encoder,
  type GetDiscriminatedUnionVariant,
  type GetDiscriminatedUnionVariantContent,
} from '@solana/kit';
import {
  getDelegateCreationArgsDecoder,
  getDelegateCreationArgsEncoder,
  getDelegateMutArgsDecoder,
  getDelegateMutArgsEncoder,
  type DelegateCreationArgs,
  type DelegateCreationArgsArgs,
  type DelegateMutArgs,
  type DelegateMutArgsArgs,
} from '.';

export type DelegateCreateOrMutateArgs =
  | { __kind: 'Create'; fields: readonly [DelegateCreationArgs] }
  | { __kind: 'Mutate'; fields: readonly [DelegateMutArgs] };

export type DelegateCreateOrMutateArgsArgs =
  | { __kind: 'Create'; fields: readonly [DelegateCreationArgsArgs] }
  | { __kind: 'Mutate'; fields: readonly [DelegateMutArgsArgs] };

export function getDelegateCreateOrMutateArgsEncoder(): Encoder<DelegateCreateOrMutateArgsArgs> {
  return getDiscriminatedUnionEncoder([
    [
      'Create',
      getStructEncoder([
        ['fields', getTupleEncoder([getDelegateCreationArgsEncoder()])],
      ]),
    ],
    [
      'Mutate',
      getStructEncoder([
        ['fields', getTupleEncoder([getDelegateMutArgsEncoder()])],
      ]),
    ],
  ]);
}

export function getDelegateCreateOrMutateArgsDecoder(): Decoder<DelegateCreateOrMutateArgs> {
  return getDiscriminatedUnionDecoder([
    [
      'Create',
      getStructDecoder([
        ['fields', getTupleDecoder([getDelegateCreationArgsDecoder()])],
      ]),
    ],
    [
      'Mutate',
      getStructDecoder([
        ['fields', getTupleDecoder([getDelegateMutArgsDecoder()])],
      ]),
    ],
  ]);
}

export function getDelegateCreateOrMutateArgsCodec(): Codec<
  DelegateCreateOrMutateArgsArgs,
  DelegateCreateOrMutateArgs
> {
  return combineCodec(
    getDelegateCreateOrMutateArgsEncoder(),
    getDelegateCreateOrMutateArgsDecoder()
  );
}

// Data Enum Helpers.
export function delegateCreateOrMutateArgs(
  kind: 'Create',
  data: GetDiscriminatedUnionVariantContent<
    DelegateCreateOrMutateArgsArgs,
    '__kind',
    'Create'
  >['fields']
): GetDiscriminatedUnionVariant<
  DelegateCreateOrMutateArgsArgs,
  '__kind',
  'Create'
>;
export function delegateCreateOrMutateArgs(
  kind: 'Mutate',
  data: GetDiscriminatedUnionVariantContent<
    DelegateCreateOrMutateArgsArgs,
    '__kind',
    'Mutate'
  >['fields']
): GetDiscriminatedUnionVariant<
  DelegateCreateOrMutateArgsArgs,
  '__kind',
  'Mutate'
>;
export function delegateCreateOrMutateArgs<
  K extends DelegateCreateOrMutateArgsArgs['__kind'],
  Data,
>(kind: K, data?: Data) {
  return Array.isArray(data)
    ? { __kind: kind, fields: data }
    : { __kind: kind, ...(data ?? {}) };
}

export function isDelegateCreateOrMutateArgs<
  K extends DelegateCreateOrMutateArgs['__kind'],
>(
  kind: K,
  value: DelegateCreateOrMutateArgs
): value is DelegateCreateOrMutateArgs & { __kind: K } {
  return value.__kind === kind;
}